the logic of user profile is breaked into separate components
