+++ client/src/components/Navbaruser.jsx
@@ -76,7 +76,7 @@ function Navbaruser() {
                 Jobs
               </Link>
             </li>
-            <li>
+            {/* <li>
               <Link
                 to="/login"
                 className="block py-2 px-3 text-gray-900 rounded-sm hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-orange-700 md:p-0 dark:text-white md:dark:hover:text-orange-500"
@@ -91,7 +91,7 @@ function Navbaruser() {
               >
                 Signup
               </Link>
-            </li>
+            </li> */}
             <li>
               <div className="relative">
                 <button
